financial-health-backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pandas
openai
cryptography
python-dotenv
pydantic
passlib[bcrypt]
python-multipart
